Moved time series point manipulation selectors into drawingData. Updated usages and start to update timeSeries.spec.js
