+++ assets/src/scripts/components/hydrograph/timeseries.js
@@ -6,33 +6,11 @@ const { createSelector } = require('reselect');
 // Create a time formatting function from D3's timeFormat
 const formatTime = timeFormat('%c %Z');
 
-export const MASK_DESC = {
-    ice: 'Ice Affected',
-    fld: 'Flood',
-    bkw: 'Backwater',
-    zfl: 'Zeroflow',
-    dry: 'Dry',
-    ssn: 'Seasonal',
-    pr: 'Partial Record',
-    rat: 'Rating Development',
-    eqp: 'Equipment Malfunction',
-    mnt: 'Maintenance',
-    dis: 'Discontinued',
-    tst: 'Test',
-    pmp: 'Pump',
-    '***': 'Unavailable'
-};
-
-// Lines will be split if the difference exceeds 72 minutes.
-export const MAX_LINE_POINT_GAP = 60 * 1000 * 72;
 
 /**
  * @return {Object}     Mapping of variable IDs to variable details
  */
-export const variablesSelector = createSelector(
-    state => state.series.variables,
-    (variables) => variables
-);
+export const variablesSelector = state => state.series.variables ? state.series.variables : null;
 
 
 /**
@@ -50,6 +28,7 @@ export const currentVariableSelector = createSelector(
  * Returns currently selected parameter code
  * @return String or null if none
  */
+    //TODO: see if this is use anywhere before removing.
 export const currentParmCd = createSelector(
     currentVariableSelector,
     (currentVar) => currentVar && currentVar.variableCode ? currentVar.variableCode.value : null
@@ -96,12 +75,15 @@ export const currentVariableTimeSeriesSelector = memoize(tsKey => createSelector
     currentVariableSelector,
     (timeSeries, variable) => {
         let ts = {};
-        Object.keys(timeSeries).forEach(key => {
-            const series = timeSeries[key];
-            if (series.tsKey === tsKey && series.variable === variable.oid) {
-                ts[key] = series;
-            }
-        });
+        if (variable) {
+            ts = {};
+            Object.keys(timeSeries).forEach(key => {
+                const series = timeSeries[key];
+                if (series.tsKey === tsKey && series.variable === variable.oid) {
+                    ts[key] = series;
+                }
+            });
+        }
         return ts;
     }
 ));
@@ -111,7 +93,7 @@ export const currentVariableTimeSeriesSelector = memoize(tsKey => createSelector
  * Selects all time series data.
  * @param  {String} tsKey   Time-series key
  * @param  {Object} state   Redux state
- * @return {Object}         Time-series data
+ * @return {Object} - Keys are tsID, values are time-series data
  */
 export const timeSeriesSelector = memoize(tsKey => createSelector(
     allTimeSeriesSelector,
@@ -127,143 +109,6 @@ export const timeSeriesSelector = memoize(tsKey => createSelector(
     }
 ));
 
-export const HASH_ID = {
-    current: 'hash-45',
-    compare: 'hash-135'
-};
-
-
-/*
- * @param {Array} points - Array of point objects
- */
-const transformToCumulative = function(points) {
-    let accumulatedValue = 0;
-    return points.map((point) => {
-        let result = {...point};
-        if (point.value !== null) {
-            accumulatedValue += point.value;
-            result.value = accumulatedValue;
-        } else {
-            accumulatedValue = 0;
-        }
-        return result;
-    });
-};
-
-export const allPointsSelector = createSelector(
-    allTimeSeriesSelector,
-    state => state.series.variables,
-    (timeSeries, variables) => {
-        let allPoints = {};
-        Object.keys(timeSeries).forEach((tsId) => {
-            const ts = timeSeries[tsId];
-            const variableId = ts.variable;
-            const parmCd = variables[variableId].variableCode.value;
-            if (ts.tsKey !== 'median' && parmCd === '00045') {
-                allPoints[tsId] = transformToCumulative(ts.points);
-            } else {
-                allPoints[tsId] = ts.points;
-            }
-        });
-        return allPoints;
-    }
-);
-
-/*
- * @param {Object} state
- * @param {String} tsKey
- * @returns {Object} of keys are tsId, values are Array of points
- * in tsKey
- */
-export const pointsByTsKeySelector = memoize(tsKey => createSelector(
-    allPointsSelector,
-    state => state.series.timeSeries,
-    (points, timeSeries) => {
-        let result = {};
-        Object.keys(points).forEach((tsId) => {
-            if (timeSeries[tsId].tsKey === tsKey) {
-                result[tsId] = points[tsId];
-            }
-        });
-        return result;
-    }));
-
-/*
-* @return Array of Array of points
- */
-export const currentVariablePointsSelector = memoize(tsKey => createSelector(
-    pointsByTsKeySelector(tsKey),
-    currentVariableTimeSeriesSelector(tsKey),
-    (points, timeSeries) => {
-        return Object.keys(timeSeries).map((tsId) => {
-            return points[tsId];
-        });
-    }
-));
-/**
- * Returns a selector that, for a given tsKey:
- * Returns an array of time points for all visible time series.
- * @param  {Object} state     Redux store
- * @param  {String} tsKey     Timeseries key
- * @return {Array}            Array of array of points.
- */
-    //TODO: Candidate for removal
-export const pointsSelector = memoize((tsKey) => createSelector(
-    pointsByTsKeySelector(tsKey),
-    (points) => {
-        return Object.values(points);
-    }
-));
-
-
-/**
- * Factory function that returns a selector for a given tsKey, that:
- * Returns a single array of all points.
- * @param  {Object} state       Redux state
- * @return {Array}              Array of points.
- */
-export const flatPointsSelector = memoize(tsKey => createSelector(
-    pointsSelector(tsKey),
-    tsPointsList => tsPointsList.reduce((finalPoints, points) => {
-        Array.prototype.push.apply(finalPoints, points);
-        return finalPoints;
-    }, [])
-));
-
-
-export const classesForPoint = point => {
-    return {
-        approved: point.qualifiers.indexOf('A') > -1,
-        estimated: point.qualifiers.indexOf('E') > -1
-    };
-};
-
-
-/**
- * Returns an array of points for each visible timeseries.
- * @param  {Object} state     Redux store
- * @return {Array}            Array of point arrays.
- */
-export const visiblePointsSelector = createSelector(
-    currentVariablePointsSelector('current'),
-    currentVariablePointsSelector('compare'),
-    currentVariablePointsSelector('median'),
-    (state) => state.showSeries,
-    (current, compare, median, showSeries) => {
-        const pointArray = [];
-        if (showSeries['current']) {
-            Array.prototype.push.apply(pointArray, current);
-
-        }
-        if (showSeries['compare']) {
-            Array.prototype.push.apply(pointArray, compare);
-        }
-        if (showSeries['median']) {
-            Array.prototype.push.apply(pointArray, median);
-        }
-        return pointArray;
-    }
-);
 
 
 /**
@@ -277,207 +122,6 @@ export const isVisibleSelector = memoize(tsKey => (state) => {
     return state.showSeries[tsKey];
 });
 
-
-/**
- * Factory function creates a function that, for a given tsKey:
- * Returns all point data as an array of [value, time, qualifiers].
- * @param {Object} state - Redux store
- * @param {String} tsKey - timeseries key
- * @param {Array of Array} for each point returns [value, time, qualifiers] or empty array.
- */
-    //TODO: This needs to be changes to use point selector since it should be showing the data being displayed
-export const pointsTableDataSelector = memoize(tsKey => createSelector(
-    allTimeSeriesSelector,
-    (timeSeries) => {
-        return Object.keys(timeSeries).reduce((dataByTsID, tsID) => {
-            const series = timeSeries[tsID];
-            if (series.tsKey !== tsKey) {
-                return dataByTsID;
-            }
-
-            dataByTsID[tsID] = series.points.map((value) => {
-                return [
-                    value.value || '',
-                    value.dateTime || '',
-                    value.qualifiers && value.qualifiers.length > 0 ? value.qualifiers.join(', ') : ''
-                ];
-            });
-
-            return dataByTsID;
-        }, {});
-    }
-));
-
-const getLineClasses = function(pt) {
-    let dataMask = null;
-    if (pt.value === null) {
-        let qualifiers = new Set(pt.qualifiers.map(q => q.toLowerCase()));
-
-        // current business rules specify that a particular data point
-        // will only have at most one masking qualifier
-        let maskIntersection = Object.keys(MASK_DESC).filter(x => qualifiers.has(x));
-        dataMask = maskIntersection[0];
-    }
-    return {
-        ...classesForPoint(pt),
-        dataMask
-    };
-};
-
-const cumulativeLineSegments = function(series) {
-    const points = series.points;
-    // Accumulate data into line groups, splitting on the estimated and
-    // approval status.
-    const lines = [];
-    let lastClasses = {};
-    let accumulatedValue = 0;
-
-    for (const pt of points) {
-        // Classes to put on the line with this point.
-        let lineClasses = getLineClasses(pt);
-        if (lineClasses.dataMask) {
-            accumulatedValue = 0;
-        }
-        if (pt.value) {
-            accumulatedValue += pt.value;
-        }
-        let newPoint = {...pt, value: accumulatedValue};
-
-        // If this point doesn't have the same classes as the last point,
-        // create a new line for it.
-        if (lastClasses.approved !== lineClasses.approved ||
-            lastClasses.estimated !== lineClasses.estimated ||
-            lastClasses.dataMask !== lineClasses.dataMask) {
-            lines.push({
-                classes: lineClasses,
-                points: []
-            });
-        }
-
-        // Add this point to the current line.
-        lines[lines.length - 1].points.push(newPoint);
-
-        // Cache the classes for the next loop iteration.
-        lastClasses = lineClasses;
-    }
-    return lines;
-};
-
-
-/**
- * Factory function creates a function that:
- * Returns all points in a timeseries grouped into line segments, for each time series.
- * @param  {Object} state     Redux store
- * @param  {String} tsKey Timeseries key
- * @return {Array}            Array of array of points.
- */
-export const lineSegmentsSelector = memoize(tsKey => createSelector(
-    pointsByTsKeySelector(tsKey),
-    (tsPoints) => {
-        let seriesLines = {};
-        Object.keys(tsPoints).forEach((tsId) => {
-            const points = tsPoints[tsId]
-            let lines = [];
-
-            // Accumulate data into line groups, splitting on the estimated and
-            // approval status.
-            let lastClasses = {};
-
-            for (let pt of points) {
-                // Classes to put on the line with this point.
-                let lineClasses = getLineClasses(pt);
-
-                // If this is a non-masked data point, split lines if the gap
-                // from the period point exceeds MAX_LINE_POINT_GAP.
-                let splitOnGap = false;
-                if (!lineClasses.dataMask && lines.length > 0) {
-                    const lastPoints = lines[lines.length - 1].points;
-                    const lastPtDateTime = lastPoints[lastPoints.length - 1].dateTime;
-                    if (pt.dateTime - lastPtDateTime > MAX_LINE_POINT_GAP) {
-                        splitOnGap = true;
-                    }
-                }
-
-                // If this point doesn't have the same classes as the last point,
-                // create a new line for it.
-                if (lastClasses.approved !== lineClasses.approved ||
-                    lastClasses.estimated !== lineClasses.estimated ||
-                    lastClasses.dataMask !== lineClasses.dataMask ||
-                    splitOnGap) {
-                    lines.push({
-                        classes: lineClasses,
-                        points: []
-                    });
-                }
-
-                // Add this point to the current line.
-                lines[lines.length - 1].points.push(pt);
-
-                // Cache the classes for the next loop iteration.
-                lastClasses = lineClasses;
-            }
-            seriesLines[tsId] = lines;
-        });
-        return seriesLines;
-    }
-));
-
-
-/**
- * Factory function creates a function that, for a given tsKey:
- * @return {Object}     Mapping of parameter to code list of line segments arrays.
- */
-export const lineSegmentsByParmCdSelector = memoize(tsKey => createSelector(
-    lineSegmentsSelector(tsKey),
-    timeSeriesSelector(tsKey),
-    variablesSelector,
-    (lineSegmentsBySeriesID, timeSeriesMap, variables) => {
-        return Object.keys(lineSegmentsBySeriesID).reduce((byVarID, sID) => {
-            const series = timeSeriesMap[sID];
-            const parmCd = variables[series.variable].variableCode.value;
-            byVarID[parmCd] = byVarID[parmCd] || [];
-            byVarID[parmCd].push(lineSegmentsBySeriesID[sID]);
-            return byVarID;
-        }, {});
-    }
-));
-
-
-/**
- * Factory function creates a function that, for a given tsKey:
- * Returns mapping of time series IDs to series for the current variable.
- * @return {Object}
- */
-const currentVariableTimeseriesSelector = memoize(tsKey => createSelector(
-    timeSeriesSelector(tsKey),
-    currentVariableSelector,
-    (seriesMap, variable) => {
-        return Object.keys(seriesMap).filter(
-                sID => seriesMap[sID].variable === variable.oid).reduce((curMap, sID) => {
-            curMap[sID] = seriesMap[sID];
-            return curMap;
-        }, {});
-    }
-));
-
-
-/**
- * Factory function creates a function that, for a given tsKey:
- * Returns mapping of series ID to line segments for the currently selected variable.
- * @return {Object}
- */
-export const currentVariableLineSegmentsSelector = memoize(tsKey => createSelector(
-    currentVariableTimeseriesSelector(tsKey),
-    lineSegmentsSelector(tsKey),
-    (seriesMap, linesMap) => {
-        const result = Object.keys(seriesMap).reduce((visMap, sID) => {
-                visMap[sID] = linesMap[sID];
-                return visMap;
-            }, {});
-        return result;
-
-    }
-));
 
 
 /**
@@ -500,7 +144,7 @@ export const titleSelector = createSelector(
 
 /**
  * @return {String}     Description for the currently display set of time
- *                      series, used by addSVGAccessibility.
+ *                      series
  */
 export const descriptionSelector = createSelector(
     currentVariableSelector,
