+++ assets/src/scripts/components/hydrograph/timeseries.spec.js
@@ -1,11 +1,12 @@
-const { lineSegmentsSelector, pointsSelector,
-    currentVariableTimeSeriesSelector, pointsTableDataSelector, allTimeSeriesSelector, MAX_LINE_POINT_GAP } = require('./timeseries');
+const { variablesSelector, currentVariableSelector, timeSeriesSelector, isVisibleSelector,
+    currentVariableTimeSeriesSelector, allTimeSeriesSelector } = require('./timeseries');
 
 
 const TEST_DATA = {
     series: {
         timeSeries: {
             '00060': {
+                tsKey: 'current',
                 points: [{
                     value: 10,
                     qualifiers: ['P'],
@@ -22,7 +23,25 @@ const TEST_DATA = {
                     approved: false,
                     estimated: false
                 }]
-            }
+            },
+            '000010': {
+                tsKey: 'compare',
+                points: [{
+                    value: 1,
+                    qualifiers: ['P'],
+                    approved: false,
+                    estimated: false
+                }, {
+                    value: 2,
+                    qualifiers: ['P'],
+                    approved: false,
+                    estimated: false
+                }, {
+                    value: 3,
+                    qualifiers: ['P'],
+                    approved: false,
+                    estimated: false
+                }]}
         },
         timeSeriesCollections: {
             'coll1': {
@@ -45,9 +64,67 @@ const TEST_DATA = {
     currentVariableID: '45807197'
 };
 
-describe('Timeseries module', () => {
+fdescribe('Timeseries module', () => {
 
-    describe('all time series selector', () => {
+    const TEST_VARIABLES = {
+            '45807042': {
+                'variableCode': {
+                    value: '00010'
+                },
+                'variableName': 'Temperature'
+            },
+            '45807197': {
+                'variableCode': {
+                    value: '00060'
+                },
+                'variableName': 'Streamflow'
+            }
+        };
+
+    describe('variablesSelector', () => {
+
+        it('should return the variables object', () => {
+            expect(variablesSelector({
+                series: {
+                    variables: TEST_VARIABLES
+                }
+            })).toEqual(TEST_VARIABLES);
+        });
+
+        it('Should return null if no variables are in the state', () => {
+            expect(variablesSelector({
+                series: {}
+            })).toBeNull();
+        });
+    });
+
+    describe('currentVariableSelector', () => {
+
+        it('should return the selected variable information', () => {
+            expect(currentVariableSelector({
+                series: {
+                    variables: TEST_VARIABLES
+                },
+                currentVariableID: '45807197'
+            })).toEqual({
+                'variableCode': {
+                    value: '00060'
+                },
+                'variableName': 'Streamflow'
+            });
+        });
+
+        it('should return null if no currentVariableID set', () => {
+            expect(currentVariableSelector({
+                series: {
+                    variables: TEST_VARIABLES
+                },
+                currentVariableID: null
+            })).toBeNull();
+        });
+    });
+
+    describe('allTimesSeriesSelector', () => {
 
         it('should return all timeseries if they have data points', () => {
             expect(allTimeSeriesSelector({
@@ -87,344 +164,6 @@ describe('Timeseries module', () => {
                 '00010': {
                     points: [1, 2, 3, 4]
                 }
-            });
-        });
-    });
-
-    describe('line segment selector', () => {
-        it('should separate on approved', () => {
-            expect(lineSegmentsSelector('current')({
-                ...TEST_DATA,
-                series: {
-                    ...TEST_DATA.series,
-                    timeSeries: {
-                        ...TEST_DATA.series.timeSeries,
-                        '00060': {
-                            points: [{
-                                value: 10,
-                                qualifiers: []
-                            }, {
-                                value: 10,
-                                qualifiers: ['A']
-                            }, {
-                                value: 10,
-                                qualifiers: ['A']
-                            }],
-                            tsKey: 'current'
-                        }
-                    }
-                }
-            })).toEqual({
-                '00060': [
-                    {
-                        'classes': {
-                            'approved': false,
-                            'estimated': false,
-                            'dataMask': null
-                        },
-                        'points': [
-                            {
-                                'value': 10,
-                                'qualifiers': []
-                            }
-                        ]
-                    },
-                    {
-                        'classes': {
-                            'approved': true,
-                            'estimated': false,
-                            'dataMask': null
-                        },
-                        'points': [
-                            {
-                                'value': 10,
-                                'qualifiers': [
-                                    'A'
-                                ]
-                            },
-                            {
-                                'value': 10,
-                                'qualifiers': [
-                                    'A'
-                                ]
-                            }
-                        ]
-                    }
-                ]
-            });
-        });
-
-        it('should separate on estimated', () => {
-            expect(lineSegmentsSelector('current')({
-                ...TEST_DATA,
-                series: {
-                    ...TEST_DATA.series,
-                    timeSeries: {
-                        ...TEST_DATA.series.timeSeries,
-                        '00060': {
-                            points: [{
-                                value: 10,
-                                qualifiers: ['P']
-                            }, {
-                                value: 10,
-                                qualifiers: ['P', 'E']
-                            }, {
-                                value: 10,
-                                qualifiers: ['P', 'E']
-                            }],
-                            tsKey: 'current'
-                        }
-                    }
-                }
-            })).toEqual({
-                '00060': [
-                    {
-                        'classes': {
-                            'approved': false,
-                            'estimated': false,
-                            'dataMask': null
-                        },
-                        'points': [
-                            {
-                                'value': 10,
-                                'qualifiers': [
-                                    'P'
-                                ]
-                            }
-                        ]
-                    },
-                    {
-                        'classes': {
-                            'approved': false,
-                            'estimated': true,
-                            'dataMask': null
-                        },
-                        'points': [
-                            {
-                                'value': 10,
-                                'qualifiers': [
-                                    'P',
-                                    'E'
-                                ]
-                            },
-                            {
-                                'value': 10,
-                                'qualifiers': [
-                                    'P',
-                                    'E'
-                                ]
-                            }
-                        ]
-                    }
-                ]
-            });
-        });
-
-        it('should separate out masked values', () => {
-            expect(lineSegmentsSelector('current')({
-                ...TEST_DATA,
-                series: {
-                    ...TEST_DATA.series,
-                    timeSeries: {
-                        ...TEST_DATA.series.timeSeries,
-                        '00060': {
-                            points: [{
-                                value: 10,
-                                qualifiers: ['P']
-                            }, {
-                                value: null,
-                                qualifiers: ['P', 'ICE']
-                            }, {
-                                value: null,
-                                qualifiers: ['P', 'FLD']
-                            }],
-                            tsKey: 'current'
-                        }
-                    }
-                }
-            })).toEqual({
-                '00060': [
-                    {
-                        'classes': {
-                            'approved': false,
-                            'estimated': false,
-                            'dataMask': null
-                        },
-                        'points': [
-                            {
-                                'value': 10,
-                                'qualifiers': [
-                                    'P'
-                                ]
-                            }
-                        ]
-                    },
-                    {
-                        'classes': {
-                            'approved': false,
-                            'estimated': false,
-                            'dataMask': 'ice'
-                        },
-                        'points': [
-                            {
-                                'value': null,
-                                'qualifiers': [
-                                    'P',
-                                    'ICE'
-                                ]
-                            }
-                        ]
-                    },
-                    {
-                        'classes': {
-                            'approved': false,
-                            'estimated': false,
-                            'dataMask': 'fld'
-                        },
-                        'points': [
-                            {
-                                'value': null,
-                                'qualifiers': [
-                                    'P',
-                                    'FLD'
-                                ]
-                            }
-                        ]
-                    }
-                ]});
-        });
-
-        it('should separate on gaps greater than MAX_LINE_POINT_GAP', () => {
-            const dates = [
-                new Date(0),
-                new Date(MAX_LINE_POINT_GAP - 1),
-                new Date(MAX_LINE_POINT_GAP),
-                new Date(2 * MAX_LINE_POINT_GAP),
-                new Date(3 * MAX_LINE_POINT_GAP + 1),
-                new Date(3 * MAX_LINE_POINT_GAP + 2)
-            ];
-            expect(lineSegmentsSelector('current')({
-                ...TEST_DATA,
-                series: {
-                    ...TEST_DATA.series,
-                    timeSeries: {
-                        ...TEST_DATA.series.timeSeries,
-                        '00060': {
-                            points: dates.map(d => {
-                                return {
-                                    value: 10,
-                                    dateTime: d,
-                                    qualifiers: ['P']
-                                };
-                            }),
-                            tsKey: 'current'
-                        }
-                    }
-                }
-            })).toEqual({
-                '00060': [{
-                    'classes': {
-                        'approved': false,
-                        'estimated': false,
-                        'dataMask': null
-                    },
-                    'points': [{
-                            'value': 10,
-                            'dateTime': dates[0],
-                            'qualifiers': ['P']
-                        }, {
-                            'value': 10,
-                            'dateTime': dates[1],
-                            'qualifiers': ['P']
-                        }, {
-                            'value': 10,
-                            'dateTime': dates[2],
-                            'qualifiers': ['P']
-                        }, {
-                            'value': 10,
-                            'dateTime': dates[3],
-                            'qualifiers': ['P']
-                        }
-                    ]
-                }, {
-                    'classes': {
-                        'approved': false,
-                        'estimated': false,
-                        'dataMask': null
-                    },
-                    'points': [{
-                        'value': 10,
-                        'dateTime': dates[4],
-                        'qualifiers': ['P']
-                    }, {
-                        'value': 10,
-                        'dateTime': dates[5],
-                        'qualifiers': ['P']
-                    }]
-                }]
-            });
-        });
-
-        it('should not separate on gaps greater than MAX_LINE_POINT_GAP if points masked', () => {
-            const dates = [
-                new Date(0),
-                new Date(MAX_LINE_POINT_GAP - 1),
-                new Date(MAX_LINE_POINT_GAP),
-                new Date(2 * MAX_LINE_POINT_GAP),
-                new Date(3 * MAX_LINE_POINT_GAP + 1),
-                new Date(3 * MAX_LINE_POINT_GAP + 2)
-            ];
-            expect(lineSegmentsSelector('current')({
-                ...TEST_DATA,
-                series: {
-                    ...TEST_DATA.series,
-                    timeSeries: {
-                        ...TEST_DATA.series.timeSeries,
-                        '00060': {
-                            points: dates.map(d => {
-                                return {
-                                    value: null,
-                                    dateTime: d,
-                                    qualifiers: ['Ice']
-                                };
-                            }),
-                            tsKey: 'current'
-                        }
-                    }
-                }
-            })).toEqual({
-                '00060': [{
-                    'classes': {
-                        'approved': false,
-                        'estimated': false,
-                        'dataMask': 'ice'
-                    },
-                    'points': [{
-                        'value': null,
-                        'dateTime': dates[0],
-                        'qualifiers': ['Ice']
-                    }, {
-                        'value': null,
-                        'dateTime': dates[1],
-                        'qualifiers': ['Ice']
-                    }, {
-                        'value': null,
-                        'dateTime': dates[2],
-                        'qualifiers': ['Ice']
-                    }, {
-                        'value': null,
-                        'dateTime': dates[3],
-                        'qualifiers': ['Ice']
-                    }, {
-                        'value': null,
-                        'dateTime': dates[4],
-                        'qualifiers': ['Ice']
-                    }, {
-                        'value': null,
-                        'dateTime': dates[5],
-                        'qualifiers': ['Ice']
-                    }]
-                }]
             });
         });
     });
@@ -507,101 +246,59 @@ describe('Timeseries module', () => {
                 four: {item: 'four', points: [4, 5], tsKey: 'current', variable: 45807197}
             });
         });
-    });
 
-    describe('pointsSelector', () => {
-        it('works with a single collection and two time series', () => {
-            expect(pointsSelector('current')({
-                series: {
-                    requests: {
-                        current: {
-                            timeSeriesCollections: ['coll1']
-                        }
-                    },
-                    timeSeriesCollections: {
-                        'coll1': {
-                            variable: 45807197,
-                            timeSeries: ['one', 'two']
-                        }
-                    },
-                    timeSeries: {
-                        one: {
-                            points: ['ptOne', 'ptTwo', 'ptThree'],
-                            tsKey: 'current',
-                            variable: 45807197
-                        },
-                        two: {
-                            points: ['ptOne2', 'ptTwo2', 'ptThree2'],
-                            tsKey: 'current',
-                            variable: 45807197
-                        }
-                    },
-                    variables: {
-                        '45807197': {
-                            variableCode: '00060',
-                            oid: 45807197
-                        }
-                    }
-                },
-                currentVariableID: '45807197'
-            })).toEqual([['ptOne', 'ptTwo', 'ptThree'], ['ptOne2', 'ptTwo2', 'ptThree2']]);
+        it('returns {} if there is no currentVariableId', () => {
+            expect(currentVariableTimeSeriesSelector('current')({
+                series: {},
+                currentVariableID: null
+            })).toEqual({});
         });
     });
 
-    describe('pointsTableDataSelect', () => {
-        it('Return an array of arrays if series is visible', () => {
-            expect(pointsTableDataSelector('current')({
-                series: {
-                    requests: {
-                        current: {
-                            timeSeriesCollections: ['coll1']
-                        }
-                    },
-                    timeSeriesCollections: {
-                        'coll1': {
-                            variable: 45807197,
-                            timeSeries: ['one']
-                        }
-                    },
-                    timeSeries: {
-                        one: {
-                            tsKey: 'current',
-                            points: [{
-                                dateTime: '2018-01-01',
-                                qualifiers: ['P'],
-                                approved: false,
-                                estimated: false
-                            }, {
-                                value: 15,
-                                approved: false,
-                                estimated: true
-                            }, {
-                                value: 10,
-                                dateTime: '2018-01-03',
-                                qualifiers: ['P', 'Ice'],
-                                approved: false,
-                                estimated: true
-                            }]
-                        }
-                    },
-                    variables: {
-                        '45807197': {
-                            variableCode: '00060',
-                            oid: 45807197
-                        }
-                    }
-                },
-                currentVariableID: '45807197',
-                showSeries: {
-                    current: true
+    describe('timeSeriesSelector', () => {
+
+        it('should return the selected time series', () => {
+            expect(timeSeriesSelector('current')(TEST_DATA)).toEqual({
+                '00060': {
+                    tsKey: 'current',
+                    points: [{
+                        value: 10,
+                        qualifiers: ['P'],
+                        approved: false,
+                        estimated: false
+                    }, {
+                        value: null,
+                        qualifiers: ['P', 'ICE'],
+                        approved: false,
+                        estimated: false
+                    }, {
+                        value: null,
+                        qualifiers: ['P', 'FLD'],
+                        approved: false,
+                        estimated: false
+                    }]
                 }
-            })).toEqual({
-                one: [
-                    ['', '2018-01-01', 'P'],
-                    [15, '', ''],
-                    [10, '2018-01-03', 'P, Ice']
-                ]
             });
+        });
+
+        it('should return null the empty set if no time series for the selected key exist', () => {
+            expect(timeSeriesSelector('median')(TEST_DATA)).toEqual({});
+        });
+    });
+
+    describe('isVisibleSelector', () => {
+        it('Returns whether the time series is visible', () => {
+            const store = {
+                showSeries: {
+                    'current': true,
+                    'compare': false,
+                    'median': true
+                }
+            };
+
+            expect(isVisibleSelector('current')(store)).toBe(true);
+            expect(isVisibleSelector('compare')(store)).toBe(false);
+            expect(isVisibleSelector('median')(store)).toBe(true);
         });
     });
 });
